msm8960: Fix for Hayabusa not booting issue
Sony Xperia TX(hayabusa) uses Simple Remote for headphone detection.
However,with the changed lines present,the phone won't boot when
Simple Remote is enabled in the kernel.And if we use MBHC Headset Control,
like Xperia T,built-in microphone will stop working.
So we must remove these lines to get the phone boot.
And it's TESTED that the lines won't affect the headphone button.
It's still able to launch and control music players.

Change-Id: I8f3c22330455e79dca9c182281eb8797f9187591